requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
prometheus-client
starlette-prometheus==0.9.0
graphene
graphene-sqlalchemy
ariadne==0.20.1
asyncio==3.4.3
statistics==1.0.3.5

asyncpg
databases[postgresql]
